Migrate to lsp4ij
