+++ build.gradle.kts
@@ -26,11 +26,9 @@ version = properties("pluginVersion")
 // Configure project's dependencies
 repositories {
     mavenCentral()
-    maven {
-        url = uri("https://jitpack.io")
-    }
     intellijPlatform {
         defaultRepositories()
+        marketplace()
     }
 }
 
@@ -42,10 +40,12 @@ dependencies {
     implementation("io.ktor:ktor-client-content-negotiation:2.2.1")
     implementation("io.ktor:ktor-serialization-kotlinx-json:2.2.1")
     implementation("org.jetbrains.kotlin:kotlin-native-utils:1.6.10")
-    implementation("com.github.ballerina-platform:lsp4intellij:0.96.1")
 
     intellijPlatform {
         create(properties("platformType"), properties("platformVersion"))
+
+        // LSP4IJ plugin dependency for Language Server Protocol support
+        plugin("com.redhat.devtools.lsp4ij:0.17.0")
 
         instrumentationTools()
         zipSigner()
