--- src/main/kotlin/com/github/zzehring/intellijjsonnet/JsonnetLSBackgroundStartupActivity.kt
@@ -1,12 +1,0 @@
-package com.github.zzehring.intellijjsonnet
-
-import com.intellij.openapi.project.Project
-import com.intellij.openapi.startup.StartupActivity
-
-class JsonnetLSBackgroundStartupActivity : StartupActivity.Background {
-    override fun runActivity(project: Project) {
-        val handler = JsonnetLSStartupHandler()
-        handler.start()
-    }
-
-}
