+++ src/main/kotlin/com/github/zzehring/intellijjsonnet/actions/EvaluateJsonnetAction.kt
@@ -1,18 +1,25 @@
 package com.github.zzehring.intellijjsonnet.actions
 
+import com.intellij.notification.Notification
+import com.intellij.notification.NotificationType
+import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.PlatformDataKeys
 import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.util.io.FileUtilRt
 import com.intellij.openapi.vfs.VfsUtil
-import org.eclipse.lsp4j.ExecuteCommandParams
+import com.redhat.devtools.lsp4ij.commands.CommandExecutor
+import com.redhat.devtools.lsp4ij.commands.LSPCommandContext
+import org.eclipse.lsp4j.Command
 import org.jetbrains.annotations.NotNull
-import org.wso2.lsp4intellij.IntellijLanguageClient
-import org.wso2.lsp4intellij.utils.FileUtils
 
 
 class EvaluateJsonnetAction : AnAction() {
+    override fun getActionUpdateThread(): ActionUpdateThread {
+        return ActionUpdateThread.BGT
+    }
+
     override fun update(@NotNull event: AnActionEvent) {
         val openedFile = event.getData(PlatformDataKeys.VIRTUAL_FILE)
         event.presentation.isEnabledAndVisible = openedFile != null && openedFile.extension == "jsonnet"
@@ -22,14 +29,40 @@ class EvaluateJsonnetAction : AnAction() {
         val tmpDir = FileUtilRt.createTempDirectory("jsonnet-plugin-tmpdir", null)
         val tmpResultFile = FileUtilRt.createTempFile(tmpDir, "jsonnet-eval", ".json")
         val openedFile = event.getData(PlatformDataKeys.VIRTUAL_FILE)
-        val project = event.project
-        val params = ExecuteCommandParams("jsonnet.evalFile", listOf(openedFile!!.path))
-        IntellijLanguageClient.getAllServerWrappersFor(FileUtils.projectToUri(project)).forEach { x ->
-            val execution = x.requestManager.executeCommand(params)
-            val result: String = execution.get() as String
-            tmpResultFile.writeText(result)
+        val project = event.project ?: return
+
+        try {
+            // Create LSP command with file path as argument
+            val command = Command("Evaluate Jsonnet File", "jsonnet.evalFile", listOf(openedFile!!.path))
+
+            // Create command context and specify our language server
+            val commandContext = LSPCommandContext(command, project)
+            commandContext.preferredLanguageServerId = "jsonnetLanguageServerId"
+
+            // Execute the command
+            CommandExecutor.executeCommand(commandContext)
+                .response()
+                ?.thenAccept { result ->
+                    if (result != null) {
+                        tmpResultFile.writeText(result.toString())
+                        val vf = VfsUtil.findFileByIoFile(tmpResultFile, true)
+                        FileEditorManager.getInstance(project).openFile(vf!!, true)
+                    }
+                }
+                ?.exceptionally { throwable ->
+                    Notification(
+                        "lsp",
+                        "Failed to evaluate Jsonnet file: ${throwable.message}",
+                        NotificationType.ERROR
+                    ).notify(project)
+                    null
+                }
+        } catch (e: Exception) {
+            Notification(
+                "lsp",
+                "Failed to evaluate Jsonnet file: ${e.message}",
+                NotificationType.ERROR
+            ).notify(project)
         }
-        val vf = VfsUtil.findFileByIoFile(tmpResultFile, true)
-        FileEditorManager.getInstance(project!!).openFile(vf!!, true)
     }
 }
