+++ src/main/resources/META-INF/plugin.xml
@@ -6,42 +6,29 @@
 
     <depends>com.intellij.modules.platform</depends>
     <depends>com.intellij.modules.lang</depends>
+    <depends>com.redhat.devtools.lsp4ij</depends>
+
+    <extensions defaultExtensionNs="com.redhat.devtools.lsp4ij">
+        <!-- Jsonnet Language Server -->
+        <server id="jsonnetLanguageServerId"
+                name="Jsonnet Language Server"
+                factoryClass="com.github.zzehring.intellijjsonnet.JsonnetLanguageServerFactory">
+            <description><![CDATA[
+Provides language support for Jsonnet files using the
+<a href="https://github.com/grafana/jsonnet-language-server">Grafana Jsonnet Language Server</a>.
+            ]]></description>
+        </server>
+
+        <!-- Map Jsonnet language to the server -->
+        <languageMapping language="Jsonnet"
+                         serverId="jsonnetLanguageServerId"
+                         languageId="jsonnet"/>
+    </extensions>
 
     <extensions defaultExtensionNs="com.intellij">
-        <!-- register a preloading activity. You need to init IntellijLanguageClient with your config, see readme -->
-        <backgroundPostStartupActivity implementation="com.github.zzehring.intellijjsonnet.JsonnetLSBackgroundStartupActivity"
-                                       id="com.github.zzehring.intellijjsonnet.JsonnetLSPreloadingActivity"/>
-
-        <!-- register intellijLanguageClient as a Service OR as a plugin component (see readme)... -->
-        <applicationService serviceImplementation="org.wso2.lsp4intellij.IntellijLanguageClient"/>
-
-        <!-- register a listener on editor events, required for lsp file sync -->
-        <editorFactoryListener implementation="org.wso2.lsp4intellij.listeners.LSPEditorListener"/>
-        <fileDocumentManagerListener implementation="org.wso2.lsp4intellij.listeners.LSPFileDocumentManagerListener"/>
-
         <!-- for displaying notifications by lsp -->
         <notificationGroup id="lsp" displayType="STICKY_BALLOON"/>
 
-        <!-- needed for completion -->
-        <completion.contributor implementationClass="org.wso2.lsp4intellij.contributors.LSPCompletionContributor"
-                                id="org.wso2.lsp4intellij.contributors.LSPCompletionContributor" language="any"/>
-
-        <!-- needed for completion as well as signature help -->
-        <typedHandler implementation="org.wso2.lsp4intellij.listeners.LSPTypedHandler"
-                      id="LSPTypedHandler"/>
-
-        <!-- needed for code diagnostics -->
-        <externalAnnotator id="LSPAnnotator" language=""
-                           implementationClass="org.wso2.lsp4intellij.contributors.annotator.LSPAnnotator"/>
-
-        <!-- needed for Workspace Symbols -->
-        <gotoSymbolContributor implementation="org.wso2.lsp4intellij.contributors.symbol.LSPSymbolContributor"
-                               id="LSPSymbolContributor"/>
-
-        <!-- needed for renaming -->
-        <renameHandler implementation="org.wso2.lsp4intellij.contributors.rename.LSPRenameHandler" id="LSPRenameHandler" order="first"/>
-        <renamePsiElementProcessor implementation="org.wso2.lsp4intellij.contributors.rename.LSPRenameProcessor" id="LSPRenameProcessor" order="first"/>
-        
         <!-- custom settings for configuring Language Server -->
         <applicationService serviceImplementation="com.github.zzehring.intellijjsonnet.settings.JLSSettingsStateComponent"/>
         <applicationConfigurable parentId="tools" instance="com.github.zzehring.intellijjsonnet.settings.JLSSettingsConfigurable"/>
@@ -59,34 +46,13 @@
         <lang.foldingBuilder language="Jsonnet"
                              implementationClass="com.github.zzehring.intellijjsonnet.JsonnetFoldingBuilder"/>
         <lang.braceMatcher language="Jsonnet" implementationClass="com.github.zzehring.intellijjsonnet.JsonnetBraceMatcher"/>
-        <formattingService implementation="com.github.zzehring.intellijjsonnet.JsonnetExternalFormatter"/>
         <lang.commenter
                 language="Jsonnet"
                 implementationClass="com.github.zzehring.intellijjsonnet.SimpleCommenter"/>
 
-        <!-- for displaying the statusbar icon -->
-        <statusBarWidgetFactory implementation="org.wso2.lsp4intellij.statusbar.LSPServerStatusWidgetFactory"
-                                id="org.wso2.lsp4intellij.statusbar.LSPServerStatusWidgetFactory"
-                                order="first" />
     </extensions>
 
-    <applicationListeners>
-        <!-- required for lsp file sync -->
-        <listener class="org.wso2.lsp4intellij.listeners.VFSListener"
-                  topic="com.intellij.openapi.vfs.VirtualFileListener"/>
-        <listener class="org.wso2.lsp4intellij.listeners.LSPProjectManagerListener"
-                  topic="com.intellij.openapi.project.ProjectManagerListener"/>
-    </applicationListeners>
-
     <actions>
-        <!-- needed for hover -->
-        <action id="org.intellij.sdk.action.QuickDocAction" class="org.wso2.lsp4intellij.actions.LSPQuickDocAction">
-        </action>
-
-        <!-- needed for find references -->
-        <action class="org.wso2.lsp4intellij.actions.LSPReferencesAction" id="LSPFindUsages">
-            <keyboard-shortcut first-keystroke="shift alt F7" keymap="$default"/>
-        </action>
         <action id="com.github.zzehring.intellijjsonnet.actions.EvaluateJsonnetAction"
                 class="com.github.zzehring.intellijjsonnet.actions.EvaluateJsonnetAction" text="Evaluate Jsonnet File"
                 description="This action evaluates the current jsonnet file ">
